Bug 1151765 - use gulp to build jsdocs, r=ricky
- use jsdoc.generator instead of gulp.dest
- use del instead of gulp-rimraf
