+++ apps/bluetooth/jsdoc.json
@@ -1,6 +1,10 @@
 {
   "bluetooth": {
-    "src": ["apps/bluetooth/js/**/*.js"],
+    "src": ["apps/bluetooth/js/**/*.js",
+            "!apps/bluetooth/js/modules/bluetooth/bluetooth_connection_manager.js",
+            "!apps/bluetooth/js/modules/bluetooth/bluetooth_context.js",
+            "!apps/bluetooth/js/modules/pair_manager.js",
+            "!apps/bluetooth/js/views/device_picker_panel.js"],
     "options": {
       "destination": "docs/bluetooth"
     }
