+++ apps/clock/jsdoc.json
@@ -1,20 +1,8 @@
 {
   "clock": {
-    "src": ["apps/clock/README.md",
-            "apps/clock/js/**/*.js",
+    "src": ["apps/clock/js/**/*.js",
             "!apps/clock/js/alarm.js",
-            "!apps/clock/js/alarm_database.js",
-            "!apps/clock/js/panels/alarm/active_alarm.js",
-            "!apps/clock/js/panels/alarm/alarm_list.js",
-            "!apps/clock/js/panels/alarm/child_window_manager.js",
-            "!apps/clock/js/panels/alarm/post_message_proxy.js",
-            "!apps/clock/js/panels/alarm_edit/main.js",
-            "!apps/clock/js/panels/timer/main.js",
-            "!apps/clock/js/picker/spinner.js",
-            "!apps/clock/js/ring_view.js",
-            "!apps/clock/js/stopwatch.js",
-            "!apps/clock/js/utils.js",
-            "!apps/clock/js/view.js"],
+            "!apps/clock/js/panels/timer/main.js"],
     "options": {
       "destination": "docs/clock"
     }
