+++ apps/fl/jsdoc.json
@@ -1,7 +1,6 @@
 {
   "fl": {
-    "src": ["apps/fl/js/**/*.js",
-            "!apps/fl/js/utils.js"],
+    "src": ["apps/fl/js/**/*.js"],
     "options": {
       "destination": "docs/forwardlock"
     }
