+++ apps/keyboard/jsdoc.json
@@ -1,9 +1,8 @@
 {
   "keyboard": {
     "src": ["apps/keyboard/js/**/*.js",
-            "!apps/keyboard/js/render.js",
-            "!apps/keyboard/js/imes/vietnamese/vietnamese.js"
-    ],
+            "!apps/keyboard/js/imes/jspinyin/libpinyin.js",
+            "!apps/keyboard/js/imes/handwriting/hwr/emscripten_zinnia.js"],
     "options": {
       "destination": "docs/keyboard"
     }
