+++ apps/operatorvariant/jsdoc.json
@@ -1,8 +1,6 @@
 {
   "operatorvariant": {
-    "src": ["apps/operatorvariant/js/**/*.js",
-            "!apps/operatorvariant/js/operator_variant.js"
-    ],
+    "src": ["apps/operatorvariant/js/**/*.js"],
     "options": {
       "destination": "docs/operatorvariant"
     }
