+++ apps/system/jsdoc.json
@@ -2,34 +2,13 @@
   "system": {
     "src": ["apps/system/js/**/*.js",
             "!apps/system/js/accessibility.js",
-            "!apps/system/js/app_chrome.js",
-            "!apps/system/js/app_install_manager.js",
-            "!apps/system/js/app_usage_metrics.js",
-            "!apps/system/js/app_window_manager.js",
-            "!apps/system/js/devtools/logshake.js",
-            "!apps/system/js/lockscreen.js",
-            "!apps/system/js/lockscreen_notification_builder.js",
-            "!apps/system/js/lockscreen_notifications.js",
-            "!apps/system/js/lockscreen_passcode_validator.js",
-            "!apps/system/js/lockscreen_state_keypadhiding.js",
-            "!apps/system/js/lockscreen_state_keypadrising.js",
-            "!apps/system/js/lockscreen_state_keypadshow.js",
-            "!apps/system/js/lockscreen_state_logger.js",
-            "!apps/system/js/lockscreen_state_manager.js",
-            "!apps/system/js/lockscreen_state_panelhide.js",
-            "!apps/system/js/lockscreen_state_slideshow.js",
-            "!apps/system/js/lockscreen_window_manager.js",
-            "!apps/system/js/migrators/browser_migrator.js",
-            "!apps/system/js/nfc_manager.js",
-            "!apps/system/js/nfc_handover_manager.js",
-            "!apps/system/js/notifications.js",
-            "!apps/system/js/places.js",
-            "!apps/system/js/shrinking_ui.js",
-            "!apps/system/js/statusbar.js",
-            "!apps/system/js/wallpaper_manager.js",
-            "!apps/system/js/browser_context_menu.js",
-            "!apps/system/js/input_layouts.js",
-            "!apps/system/js/rocketbar.js"],
+            "!apps/system/js/app_window.js",
+            "!apps/system/js/bluetooth_transfer.js",
+            "!apps/system/js/bluetooth_v2.js",
+            "!apps/system/js/browser_mixin.js",
+            "!apps/system/js/card.js",
+            "!apps/system/js/system_dialog.js",
+            "!apps/system/js/migrators/settings_migrator.js"],
     "options": {
      "destination": "docs/system"
     }
