+++ package.json
@@ -5,11 +5,13 @@
   "description": "A bundle of components which Gaia needs.",
   "author": "The Gaia Team <dev-gaia@lists.mozilla.org>",
   "devDependencies": {
-    "async": "0.2.9",
     "adm-zip": "0.4.3",
+    "archiver": "^0.10.0",
+    "async": "0.2.9",
     "chai": "1.9.1",
     "dateformat": "1.0.6-1.2.3",
     "debug": "*",
+    "del": "^1.1.1",
     "download": "0.1.7",
     "ecstatic": "0.5.4",
     "ejs": "0.8.4",
@@ -17,9 +19,9 @@
     "fe.js": "~0.1.4",
     "ffi-prebuilt": "~1.2.6",
     "gaia-raptor": "1.3.0",
-    "grunt": "0.4.2",
-    "grunt-contrib-clean": "0.5.0",
-    "grunt-jsdoc": "0.5.1",
+    "gulp": "3.8.11",
+    "gulp-babel": "^5.1.0",
+    "gulp-jsdoc": "0.1.4",
     "istanbul": "^0.3.2",
     "jshint": "2.7.0",
     "load-grunt-tasks": "0.4.0",
@@ -30,27 +32,26 @@
     "marionette-device-host": "1.0.0",
     "marionette-file-manager": "0.0.4",
     "marionette-helper": "~0.3.1",
+    "marionette-js-runner": "1.0.15",
     "marionette-plugin-forms": "0.0.3",
     "marionette-profile-builder": "0.0.4",
-    "marionette-js-runner": "1.0.15",
     "marionette-settings-api": "~0.1.0",
+    "mime": "^1.2.11",
     "mocha": "1.20.0",
     "mocha-json-proxy": "~0.0.7",
     "mocha-parallel": "0.1.0",
     "mocha-tbpl-reporter": "0.2.1",
     "moment": "~2.4.0",
     "mozilla-download": "1.0.4",
-    "proxyquire": "~0.4",
     "node-static": "0.6.9",
+    "proxyquire": "~0.4",
     "q": "~0.9.7",
     "request": "2.33.0",
     "restify": "2.8.1",
     "rimraf": "2.2.5",
+    "slugid": "^1.0.2",
     "test-agent": "0.28.2",
-    "watch": "~0.9.0",
-    "archiver": "^0.10.0",
-    "mime": "^1.2.11",
-    "slugid": "^1.0.2"
+    "watch": "~0.9.0"
   },
   "dependencies": {
     "diveSync": "^0.2.1",
